Add SpatialJoin class
To use this class, instantiate it passing it the path to a geojson file and a file containing long and lat points. Then call the method spatial_join_and_save. This will result in the points being mapped into the locations of the geojson file and save as a new csv which includes the calculated positions.